fix: use proper buffer size with BoundedStorageProvider when content length is known (#37)
Resolves #36 

Currently, `BoundedStorageProvider` prefers using the content length
when it is known. This causes problems when using a
`BoundedStorageProvider` to stream a very large file rather than an
infinite stream. We should compare the content length with the provided
buffer size and choose the smaller of the two in order to respect the
user input and prevent allocating a buffer larger than the content
length.