chore(deps): replace "futures" with "futures-util" (#176)
This PR changes `stream-download` to depend on `futures-util` directly
as `futures` includes (including dependencies) a lot more than is
actually used in this crate.
It *might* be possible thin it down even more, but this should reduce it
at least a bit already.

PS: this does not entirely drop `futures` from the workspace due to some
dev-dependencies using it, but it will for consumers of
`stream-download`